… app/schemas/com.duckduckgo.app.global.db.AppDatabase/23.json
… app/src/androidTest/java/com/duckduckgo/app/onboarding/store/AppUserStageStoreTest.kt
… app/src/androidTest/java/com/duckduckgo/app/statistics/VariantManagerTest.kt
… app/src/androidTest/java/com/duckduckgo/app/systemsearch/SystemSearchViewModelTest.kt
… app/src/androidTest/java/com/duckduckgo/app/tabs/model/TabDataRepositoryTest.kt
… app/src/main/java/com/duckduckgo/app/browser/BrowserTabViewModel.kt
… app/src/main/java/com/duckduckgo/app/browser/BrowserViewModel.kt
… app/src/main/java/com/duckduckgo/app/browser/shortcut/ShortcutBuilder.kt
… app/src/main/java/com/duckduckgo/app/browser/shortcut/ShortcutReceiver.kt
… app/src/main/java/com/duckduckgo/app/di/DatabaseModule.kt
… app/src/main/java/com/duckduckgo/app/di/NotificationModule.kt
… app/src/main/java/com/duckduckgo/app/global/DuckDuckGoApplication.kt
… app/src/main/java/com/duckduckgo/app/global/ViewModelFactory.kt
… app/src/main/java/com/duckduckgo/app/global/db/AppDatabase.kt
… app/src/main/java/com/duckduckgo/app/global/useourapp/UseOurAppDetector.kt
… app/src/main/java/com/duckduckgo/app/notification/AndroidNotificationScheduler.kt
… app/src/main/java/com/duckduckgo/app/notification/model/UseOurAppNotification.kt
… app/src/main/java/com/duckduckgo/app/onboarding/store/UserStageStore.kt
… app/src/main/java/com/duckduckgo/app/statistics/VariantManager.kt
… app/src/main/java/com/duckduckgo/app/tabs/db/TabsDao.kt
… app/src/main/java/com/duckduckgo/app/tabs/model/TabDataRepository.kt