<!--
Note: This checklist is a reminder of our shared engineering expectations. Feel free to change it, although assigning a GitHub reviewer and the items in bold are required.
-->

Task/Issue URL: https://app.asana.com/0/1125189844152671/1186436893520548
Tech Design URL: 
CC: 

**Description**:
This PR prepares the in app usage flow for the second experiment, includes:
1. New variants.
1. Removes the migration to move old user into the new in app flow.
1. New pixel when a user visits the use our app URL before the initial notification.
1. Removes CTA checks before firing some pixels to be consistent with the control group.
1. Removes day 1 and 3 notifications and onboarding on day 3 in 2 of the new variants.
1. When opening a shortcut we now try to match the domain rather than the full URL only for use our app URLs.
1. Sets to 0 weight on old variants.
1. Removes code around old notifications used on a previous experiment for more readability purposes.

**Before testing**
In `DuckDuckGoApplication` lines 221 and 222 change the code to:

```
val days = TimeUnit.MILLISECONDS.toSeconds(System.currentTimeMillis() - appInstallStore.installTimestamp)
if (days >= 30) {
```

You will be now moved out of the onboarding stage after 30 seconds from install whenever you resume the app.

In `AndroidNotificationScheduler` lines 58 and 59 change the code to be `40` and `TimeUnit.SECONDS`. Also change the line 75 to be:

```
scheduleNotification(OneTimeWorkRequestBuilder<PrivacyNotificationWorker>(), 40, TimeUnit.SECONDS, UNUSED_APP_WORK_REQUEST_TAG)
```

The notifications will be now triggered after 40 seconds.

### **Steps to test this PR**:

#### **Updating from previous version when in use our app notification stage**
1. Install from develop.
1. Launch the app and manually change the `userStage` table in the database inspector to be `USE_OUR_APP_NOTIFICATION` in the appStage column.
1. Close the app and do not relaunch, this simulates users updating to this version that could potentially be trapped in this stage.
1. Change to this bran and before updating make sure you make the changes in the before testing section.
1. Update and launch the app.
1. The `appStage` in the database should now be `ESTABLISHED`
1. Wait for 40 seconds.
1. You should receive the privacy notification.

#### **Variant zl - In App usage flow**
Hardcode the variant to use `zl`

**Clicking the notification moves you out of the onboarding**
1. Install and launch the app.
1. Go to SERP.
1. You should see the SERP CTA, press Phew.
1. Stay in that screen for at least 40 seconds.
1. The new in-app usage notification should show app.
1. Notice the day 1 privacy notification does not show up.
1. Click on the notification, you should be taken to a new tab showing the hombre screen with the new CTA.
1. Ignore the CTA (click Not now) and go to `instagram.com`.
1. You should not see any onboarding CTAs since you are not in the onboarding stage anymore.

**After X time you are moved out of the onboarding**
1. Install and launch the app.
1. Go to SERP.
1. You should see the SERP CTA, press Phew.
1. Stay in that screen for at least 30 seconds (if you see the in app usage notification do not interact with it or change the timing to a higher value so you don't see it).
1. Send the app to the background or kill the app.
1. Relaunch the app
1. Go to instagram.com
1. You should not see any onboarding CTAs since you are not in the onboarding stage anymore.

**Hiding tips does not show the notification**
1. Install and launch the app.
1. Go to SERP.
1. You should see the SERP CTA, click hide tips and hide them forever.
1. Stay in that screen for at least 40 seconds 
1. You should not see any notifications.

#### **Variant zk - Removes onboarding and day1/3 notifications**
Hardcode the variant to use `zk`

**After X time you are moved out of the onboarding and no notifications are scheduled**
1. Install and launch the app.
1. Go to SERP.
1. You should see the SERP CTA, press Phew.
1. Stay in that screen for at least 40 seconds.
1. No notifications should be shown.
1. Send the app to the background or kill the app.
1. Relaunch the app
1. Go to instagram.com
1. You should not see any onboarding CTAs since you are not in the onboarding stage anymore.

**Pixels and tabs**
1. Launch the app .
1. Go to facebook.com
1. Pixel `m_uoa_v` should be sent.
1. In the menu click Add to Home.
1. The Add to Home Screen dialog should show with a high-res icon.
1. Click in Add.
1. Pixel `m_uoa_s_a` should be sent
1. Login to facebook.
1. The fireproof dialog should not be shown.
1. Navigate to your notifications.
1. Kill the app.
1. Open the shortcut you just created.
1. Pixels `m_sho_uoa_o` and `m_uoa_v` should be sent.
1. The notifications tab you had opened before should be opened instead of a new tab.
1. Close the tab and send the app to the background.
1. Open the shortcut.
1. A new tab should be created pointing to facebook.com

#### **Variant zj - Control**
Hardcode the variant to use `zk`

**Notification and onboarding**
1. Install and launch the app.
1. Go to SERP.
1. You should see the SERP CTA, press Phew.
1. Stay in that screen for at least 40 seconds.
1. The privacy notification should be shown.
1. Send the app to the background or kill the app.
1. Relaunch the app
1. Go to instagram.com
1. You should see a CTA.

**Pixels and tabs**
1. Launch the app .
1. Go to facebook.com
1. Pixel `m_uoa_v` should be sent.
1. In the menu click Add to Home.
1. The Add to Home Screen dialog should show with a high-res icon.
1. Click in Add.
1. Pixel `m_uoa_s_a` should be sent
1. Login to facebook.
1. The fireproof dialog should not be shown.
1. Navigate to your notifications.
1. Kill the app.
1. Open the shortcut you just created.
1. Pixels `m_sho_uoa_o` and `m_uoa_v` should be sent.
1. The notifications tab you had opened before should be opened instead of a new tab.
1. Close the tab and send the app to the background.
1. Open the shortcut.
1. A new tab should be created pointing to facebook.com

---
###### Internal references:
[Software Engineering Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design Template](https://app.asana.com/0/59792373528535/184709971311943)